refactor(admin): split payments panels
Move the Quantercise payments display panels out of the route module. The page now owns environment lookup, analytics fetching, suspense fallback, and error handling while the revenue, subscriber, dunning, failed payment, dispute, refund, and MRR panels live together in a focused module.

Verified: pnpm exec prettier --check apps/admin/src/app/quantercise/payments/page.tsx apps/admin/src/app/quantercise/payments/payments-panels.tsx

Verified: git diff --check -- apps/admin/src/app/quantercise/payments/page.tsx apps/admin/src/app/quantercise/payments/payments-panels.tsx

Verified: pnpm --filter @anipotts/admin typecheck

Verified: pnpm --filter @anipotts/admin lint

Verified: pnpm --filter @anipotts/admin build

Verified: pnpm validate

Verified: pnpm dlx knip --no-exit-code
